textproc/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
num2words==0.5.13
prometheus-client==0.21.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0

